Linkedin-scrapper/requirements.txt
# Core dependencies
httpx[http2]>=0.27.0
beautifulsoup4>=4.12.2
pandas>=2.3.3
python-dotenv>=1.0.0
fake-useragent>=1.4.0
lxml>=6.0.0

# Enhanced features
tqdm>=4.66.0              # Progress bars
openpyxl>=3.1.2           # Excel export support
